Update index.html
Altered the form element so that it sends messages to my email as well as for accessibility. Also added lorem ipsum text in a paragraph to serve as a hypothetical project description so that I can experiment with the jQuery and get it to display a project's description when selected by the user.