Update scripts.js
Switched to toggling fade outs to remove other projects/descriptions when one selects/unselects a project. This has led to a bug where the changes in one layout (tablet-and-up or mobile) effects the layout of the other if one changes the screen size before setting everything to the way it was before. I'll see if I can come up with a way to get JavaScript to detect the media transition and reload the page at that time to work around this bug.