Update scripts.js
Tried to detect the small animation in the CSS as a way of detecting a breakpoint so that the page could be reloaded, but that is very buggy. Also tried a JavaScript library for detecting CSS media queries (Enquire.js), but I can't get it to work. I think I may have to just refactor the code to prevent the bug.